Add files via upload
Project made for CS50, Introduction to Computer Science. Using minimax algorithm, I made my first AI that plays (pretty well) against you. It looks five moves ahead into the game and chooses the move that most favors its win, avoiding ones that will result in its loss. As of Feb. 26, 2020, of the dozens who have played it, only two people have beat the program. 

Credits to people for providing the framework of the program is in the file.